fw-4255, move predicate utils to permissions package
