+++ firstvoices/backend/permissions/predicates/__init__.py
@@ -1,4 +1,3 @@
 from .base import *  # noqa F401
-from .utils import *  # noqa F401
 from .view import *  # noqa F401
 from .view_models import *  # noqa F401
