+++ firstvoices/backend/tests/test_predicates/__init__.py
@@ -1,6 +1,6 @@
 from . import (  # noqa F401
+    test_permission_utils,
     test_predicate_base,
-    test_predicate_utils,
     test_predicate_view,
     test_predicate_view_models,
 )
